1. Init kprocess pid_0 into struct proc and set it sp reg
2. Support to create other kprocess by fork()

3. Support to switch other kprocess by switch_to()

+++ lib/include/type.h
@@ -29,8 +29,8 @@ typedef unsigned short u16;
 typedef int s32;
 typedef unsigned int u32;
 
-typedef long long s64;
-typedef unsigned long long u64;
+typedef long s64;
+typedef unsigned long u64;
 
 #ifndef _SIZE_T
 #define _SIZE_T
